backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
torch==2.2.0
transformers==4.38.1
bitsandbytes==0.42.0
peft==0.7.1
accelerate==0.27.2
sentencepiece==0.1.99
